+++ src/main/java/nl/saxion/game/Bloodspire/Classes/Item.java
@@ -4,13 +4,13 @@ public class Item {
     String itemName;
     String itemType;
     String rarity;
-    double hitpointsValue;
-    double defenseValue;
-    double damageValue;
-    double speedPenalty;
+    int hitpointsValue;
+    int defenseValue;
+    int damageValue;
+    int speedPenalty;
     int itemID;
 
-    public Item(String itemName, String itemType, String rarity, double hitpointsValue, double defenseValue, double damageValue, double speedPenalty, int itemID) {
+    public Item(String itemName, String itemType, String rarity, int hitpointsValue, int defenseValue, int damageValue, int speedPenalty, int itemID) {
         this.itemName = itemName;
         this.itemType = itemType;
         this.rarity = rarity;
